requirements.txt
Flask==2.3.2
plaid-python==9.5.0
python-dotenv==1.0.0